+++ tests/helpers/maplibregl.js
@@ -1,0 +1,33 @@
+(function () {
+    "use strict";
+
+    window.ol = {
+        setZoom: function () {},
+        setPitch: function () {},
+        setZoom: function () {},
+        setMinZoom: function () {},
+        setMaxZoom: function () {},
+        setPitch: function () {},
+        setMinPitch: function () {},
+        setMaxPitch: function () {},
+        setCenter: function () {},
+        setStyle: function () {},
+        fitBounds: function () {},
+        on: function () {},
+        getCenter: function () {},
+        getZoom: function () {},
+        getPitch: function () {},
+        getSource: function () {},
+        getBounds: function () {},
+        addSource: function () {},
+        addLayer: function () {},
+        addControl: function () {},
+        moveLayer: function () {},
+        removeLayer: function () {},
+        removeControl: function () {},
+        rotateTo: function () {},
+        setPaintProperty: function () {},
+        flyTo: function () {},
+        Map: function () {}
+    };
+})();