+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## MapLibre GL widget v0.x.0
 
+- Prepare unit test
 - Fix animating
 
 ## MapLibre GL widget v0.4.0 - April 25, 2021