+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ## MapLibre GL widget v0.x.0
 
-- Prepare unit test
-- Fix animating
+- Support webpack
+- Prepare unit test (15)
+- Fix animating (#14)
 
 ## MapLibre GL widget v0.4.0 - April 25, 2021